+++ src/sdk/model/transaction/transaction.cpp
@@ -7,7 +7,7 @@
 namespace xpx_sdk {
 	
 	Transaction::Transaction(TransactionType type,
-	                         uint16_t fullVersion,
+	                         uint32_t fullVersion,
 	                         Amount maxFee,
 	                         int64_t deadline,
 	                         std::vector<uint8_t> binaryData,
