+++ src/sdk/model/transaction/transaction_utils.h
@@ -7,8 +7,8 @@
 using xpx_sdk::Transaction;
 namespace xpx_sdk { namespace internal {
 	
-	uint16_t MakeTransactionFullVersion(uint8_t version, NetworkIdentifier networkId);
-	uint8_t ExtractTransactionVersion(uint16_t fullVersion);
+	uint32_t MakeTransactionFullVersion(uint32_t version, NetworkIdentifier networkId);
+	uint32_t ExtractTransactionVersion(uint32_t fullVersion);
 	NetworkIdentifier ExtractTransactionNetworkId(uint16_t fullVersion);
 
 	RawBuffer GetTransactionSignedData(const Transaction* transaction);
